EASY: add missed site to check whether shutdown has been signaled
Summary:
In principle, we should check whether the shutdown has been signaled each time
after we obtain the ncm_weak_ptr.lock() which possibly extended NCM's lifetime
past that of the Processor.

This still relies on the fact that the callback of `collectAllSemiFuture` will
be executed on the last finished Worker or not executed at all--in which case,
the Processor / Worker is still alive as long as the shutdown has not been
signaled. Otherwise we would need further synchronization.

Reviewed By: MohamedBassem

Differential Revision: D14790263

fbshipit-source-id: 8bf9761947788d5d1542e4ff54673323ca41ad0e