Add the enabled flag to the SequencerAttributes & deprecate old sequencer membership format
Summary:
Let's store in the sequencer attribute if the "sequencer" flag was true or false. This is needed for the ServerConfig::NodesConfig to NodesConfiguration conversion. And with this flag stored, we can start serializing the ServerConfig using the new format.
The diff hides a weight behind a getter that hides the existence of the flag and mimics the old behavior.
This diff should theoretically be a no-op on the sequencer behavior.

This diff also drops support for the old format.

Reviewed By: scsiguy

Differential Revision: D14508260

fbshipit-source-id: be09fbf6bc730fcb03fc4e4c72afc0ea3c39c97b