ClientReadStream: log current LSN and metadata effective_since when rewinding
Summary: I saw a confusing ld-read-test-app output with repetitive messages about rewinds. It looked as if it's in some sort of rewind loop switching authoritative status back and forth, but actually it was only rewinding once every few nodeset changes, and everything was working as intended. This piece of logging would make it clearer.

Reviewed By: lucaspmelo

Differential Revision: D14418492

fbshipit-source-id: 883c6d1c6514fbea445a6656d7c20c21510afa11