set NCM in processor in NCM::init
Summary:
In testing the Logger changes, I found that one of the NCM tests was
sporadically failing under stress. The issue was that `processor->setNCM` was
not called so Processor shutdown did not invoke NCM shutdown, which caused
certain races.

This diff adds `Prosessor::setNodesConfigurationManager` in `NCM::init` to
ensure this isn't neglected.

Reviewed By: MohamedBassem

Differential Revision: D15042551

fbshipit-source-id: 1cebcc1283e833c62d6ab1c11fa1fc6d0db7457a