Log NCM update and overwrite
Summary:
This diff logs update and overwrite calls to NCM. It changes the `nc_update_`
field to be a lazily invoked function so that we don't generate the string
field unless the sample is going to be logged. It also adds a hacky
`debugJsonString` to NCCodec, so that we could log the content of the NC on
overwrite. We may want a more compact format.

Going forward, it would also be nice to have a "RequestContext" that can be
threaded through the lifetime of an NCM request, where we could collect all the
raw information (serialized and deserialized NC, etc.) to avoid redundant work
and make object lifetime easier.

Reviewed By: MohamedBassem

Differential Revision: D14988234

fbshipit-source-id: 6741dc1fbb932722904b49295cb1353b57ae0494