Throttle rebuilding based on total capacity and capacity being rebuilt
Summary:
Currently, the rebuilding throttler has two mechanisms:
1. Capacity based: A fraction of nodes are allowed to be rebuilt
2. Rack based: no more than a rack of nodes can arrive for rebuilding in a given instance

The rack check may be too conservative if the cluster is MSB balanced. In addition this logic is spread in two different places.

This change enhances the logic to not be rack size constrained. It now only allows a specified percentage of nodes to be rebuilding at a given time. In addition, in a given round, if enough triggers arrive that we may go over our threshold, then the entire queue of incoming triggers is not admitted for rebuilding. This is to ensure that if, due to a configuration or other human error, a large number of nodes fail we don't start rebuilding some of them only to reach our threshold and reject the remaining. I preserved the preexisting cool-down period when we enter throttling mode.

Draining/Drained nodes are not considered for the above throttling. This ensures that admin drain operations do not end up inadvertently throttle the rebuilding or organic failures.

Reviewed By: scsiguy

Differential Revision: D14310280

fbshipit-source-id: f300a0866b95fe949e272816de5267ba97ec3597