TimerMap interface clean up
Summary:
TimerMap is relaying on a user to give it container for output only for a case when you try to get a common timer and map itself is full.
This has lead to a weird state where EventLoop holds a static holder for a temporary struct.
If someone stores this pointer it's possible it will schedule a timer with wrong time.

Reviewed By: tau0

Differential Revision: D15080205

fbshipit-source-id: 93a14ead0b6dbff0a0dd8097b1c20023e95ab8f8