Fixing the data race in BoycottTracker
Summary:
- `isBoycotted` used to be called from multiple threads (mostly from AppendPrep) and it wasn't thread safe which made it racy. A concurrent call to `isBoyoctted` while `boycotted_nodes_.clear()` is called will lead to a crash. I changed it to be a `FastUpdateableSharedPtr`, which is a fast thread-safe shared_ptr.
- `boycotted_nodes_` was a map, but it doesn't need to, so I switched it into an unordered_set.
- For better encapsulation, I moved the `calculateBoycottedNodes` to be a private function and used the public API in the test instead.

The `FastUpdateablePointer` should be fast enough, specially that I'm doing a `compare_and_swap` which means that the local cached pointer won't change unless there's a change in the boycotted set (which is infrequent). I'm still a bit concerned about the performance though specially that this `isBoycotted` is in the critical path of appends, any ideas how to make sure this doesn't introduce a performance regression?

Reviewed By: tau0

Differential Revision: D15062391

fbshipit-source-id: 5014f74e38825c2c21b01c2409cd0ea5660a22ae