Fix for NodeAvailabilityChecker/Socket
Summary:
When we never tried to connect on a socket Socket returns that we are just in the process of connecting and NodeAvailabilityChecker declares a node as a viable copyset pick.
Send can now fail with INTERNAL sicne we fixed immediate failure and there was a check around it.
This fix makes NodeAvailabilityChecker proactively try to connect as it does in NOT_CONNECTED state and sets the proper return status.

Reviewed By: gdrane

Differential Revision: D15129880

fbshipit-source-id: 3fbb7bebdc12d6f8aa0ce0b624abee8a4c1e10a8