We now run AdminServer in non-fb builds via ThriftServer
Summary:
This diff reorganises the code a bit to allow us to create and run AdminServer objects in all builds (including non-fb). For this the following changes have been made:
- `AdminServer` abstract class now creates the `ThriftServer` object and pulls some of the common functionality in.
-  `SimpleAdminServer` is an implementation of the interface `AdminServer` that uses `apache::thrift::util::ScopedServerThread` to manage the underlying `ThriftServer` created in the abstract interface `AdminServer`
- Admin API integration tests were added and are being run against the public non-fb build.

This breaks the open source build as this is the first incident where we introduce hard dependency on thrift in OSS builds. I didn't touch the Cake files in this diff but will address the open source build in a separate diff.

Reviewed By: MohamedBassem

Differential Revision: D15165145

fbshipit-source-id: a64c8e9932b4ca0589dc6fa8f23721c6778b6127