Splitting logdevice server core code from the main runnable
Summary:
This diff moves everything related to the legacy command port into a single directory (server/admincommands/) to help with the CMake globbing we do in OSS build.

This also helps decouple the entangled dependency we have between admin commands and the rest of the server code. the core of the server should not depend on the legacy command port commands.

Reviewed By: MohamedBassem

Differential Revision: D15239713

fbshipit-source-id: f901dbe5e86cb4fb91a33f6ea856a8db2f393239