Moving LogMetaDataFetcher into admin and decoupling from Client
Summary: This is the first step needed to build a new safety checker that doesn't need to fetch the metadata for each request. This utility class replaces a large portion of the work that safety checker currently does. I'm decoupling its dependency on Client and moving it to the admin/safety while keeping it as a separate target so that tools can use it if needed.

Reviewed By: MohamedBassem

Differential Revision: D14722534

fbshipit-source-id: c22a34547cc19dab14db334eb0ee8f5164624a2a