SafetyChecker main API to use futures
Summary:
This cleans up the `SafetyChecker` object to be the primary object that I will use in future diffs. It's going to be a long-living object that holds and manages the metadata fetches and answers checkImpact queries asynchronously.

The currently  blocking checkImpact is only being used from our python bindings while every other async case sends the `CheckImpactRequest` through processor. This changes this API to be future-based so that all other cases (including the python bindings) go through this code.

The python binding will `get()` the future since it's synchronous.

Cleaner == Happier.

Reviewed By: pvjykumar

Differential Revision: D14973661

fbshipit-source-id: bb0a0e270f5c86e1977e32d234afcb1a085ea172