Close slow sockets.
Summary:
I have seen this multiple times during testing that a socket does not close for
a long time and hoards lots of bytes in it's pending queue. Not sure if others
saw. The socket is still connected but has a very high rtt . I also saw that
other workers can send to this node normally. This might be a kernel or
libevent bug. But recreating this scenario is difficult but happens frequently
on our clusters.  This leads to this worker never being able to send any data
through that socket for a very long time untill someone resets the node.
Ideally we need to have a timeout on the socket even though we do non blocking
io.

Created a timer on sender to check for slow sockets every 10s and close them if
a message was not drained for max_time_to_drain. max_time_to_drain by default is set to 15s.

This is will not only help enobufs socket but also help if socket is very slow
in draining messages.

Reviewed By: mcrnic

Differential Revision: D14493883

fbshipit-source-id: 7bf097e1f00e9e52e56181c7509f765752006188