Bug fix: Ignore acked nodes from the rebuilding set when computing rebuilding-nodes
Summary:
This is a bug fix to the original diff D14310280.

Relevant to this bug, prior to D14310280 rebuilding throttling did not account for drained/draining nodes. So it used the ShardAuthoritativeStatusMap which already filtered them out. The above diff needed to account for them so it iterated the rebuilding set directly. But as a result failed to ignore nodes that are in the rebuilding set that have already acked their rebuilding and counted them as still rebuilding. Fixed.

This bug actually hit one of the E2E systems where some of the nodes that were shrunk were not expanded back so they failed to ack their rebuilding completion. As a result even the acked nodes were counted as still rebuilding. Thanks Mohamed Bassem for discovering and triaging the issue.

Reviewed By: pvjykumar

Differential Revision: D15385720

fbshipit-source-id: 8a572f1405bc72c87bd610748896a4e92dcfe7e8