Add a new 'name' field to the NodesConfig
Summary:
This diff gives the nodes in the cluster a human readable name. Previously, we used to hack this information in the config using the `hostname` field that was used only by the tooling. But now that we'll be doing the manageability work through the admin server, we will need to formalize this field in the config. This field will also be used later on for the self registration project and it will act the node's stable identifier.

Note: This field will eventually be required, but because of config synchronization, servers will need to serialize this field first before clients can enforce it. I opened T44427489 to track this.

This diff:
- Adds a new field "name" to the ServerConfig and NodesConfiguration.
- Validates the uniqueness of the name in the config if it's set.
- Deprecates the old "hostname" field from the NodesConfiguration thrift.
- Started adding default values to the "name" field in our test and integration test utils.
- Added the name field to the NodesConfig in the Admin API.

Reviewed By: AhmedSoliman

Differential Revision: D15351713

fbshipit-source-id: c51988521265d8f27ef25763f5467058fad2bdaa