Start MaintenanceManager on server in test
Summary:
This patch creates MaintenanceManager on regular server if a setting (enable-maintenance-manager-in-test) is enabled and attaches it to the admin server.

PS: Normally we do not want to create MaintenanceManager on servers. This is done for testing purposes and the creation is controlled by a setting exclusively used in tests (I don;t want to use enable-maintenance-manager since I don't want it to be accidentally set in config which would result in all nodes starting maintenance manager)

Reviewed By: AhmedSoliman

Differential Revision: D15347029

fbshipit-source-id: c555298e6b41569cdf6b51e78c669c95ad39c410