SafetyChecker to consider the STARTING_UP state correctly on checks
Summary:
SafetyChecker is happy if nodes are in either STARTING_UP or ALIVE to consider a node to be working. This is not true for non-internal-logs. The node has to be fully started to consider it fully functional.

This diff fixes that.

The metadata logs are kept to the old behaviour because these operations don't need to wait for the logs config (metadata log configuration lives in server config) so they function without fully loaded logs config.

Reviewed By: MohamedBassem

Differential Revision: D15414987

fbshipit-source-id: fc375c35b085c317bdb17b1363e8c39fd646820a