Remove Listener inheritance from EventLoop
Summary:
This is part of changes which simplify EventLoop shutdown process.
First step is to remove all inheritances from EventLoop.
As a side effect this also allows us to run multiple listeners on the same EventLoop or on an EventLoop used for some other work

It also introduces shutdown mechanics which will be used throughout the stack and demonstrate how they should be used

Reviewed By: tau0

Differential Revision: D15361177

fbshipit-source-id: cab11e0b991eb9e31be09c6601b3e55ba4ae5da9