EASY: change NC::applyUpdate to take `const Update&`.
Summary:
`NodesConfiguration::Update` is only movable due to its `unique_ptr` members.
`NodesConfiguration::applyUpdate` does not really consume the contents of the
`Update` object, but currently takes it by value, which forces the callsite to
yield ownership. In addition, `PerRoleConfig::applyUpdate` etc. already takes
`const&`.

Conceivably the callsite would find the update useful later for logging or
other checking. In particular, for the RequestContext of `NCM::update()` (WIP),
it'd be useful to keep the `Update` objects around through the life time of the
request.

(This change does not affect existing callsites that move the update object
in, which is nice.)

Reviewed By: MohamedBassem

Differential Revision: D15471704

fbshipit-source-id: 5f5bd7edfa8ffdff789d89f9139e939335a6c16c