add NCMRequestData and NCMContext for observability
Summary:
This diff implements a mechanism to thread through context data for each NCM
query. Each NCM update or overwrite involves multiple steps and multiple
requeuing on Processor / Worker queues or RPCs. The RequestData is a good place
to accumulate debugging information (like timestamps) should the query fails /
takes unusually long. The intention is to log this to a separate error /
outlier TraceLogger. Potential future benefit could include avoiding
serialization and deserialization multiple times.

`NCMContext` is like a "scope guard" except (1) it contains `NCMRequestData`
instead of a function; and (2) it is meant to be passed around and only execute
the `RequestData::onDestruction` on the final dtor (ScopeGuard is similar but
most find it intuitive to pass it around).

It is implemented using templates instead of inheritance because one cannot
invoke virtual methods of a derived class from the base class ctor or dtor.

This diff implements NCMContext for `overwrite` and `update`.

Reviewed By: MohamedBassem

Differential Revision: D15262886

fbshipit-source-id: 0a2d74a831c75f0e34cf4ef12fa1e1963128b2d2