Fix lock-order-inversion in FileConfigSource
Summary:
This diff fixes another lock-order-inversion reported by TSAN. There are two locks that are acquired in different order depending on the method that gets called.
1/ Synchronized TextConfigUpdaterImpl's lock
2/ FileConfigSource's mutex

getConfig is called from synchronized object (meaning, 1 is acquired) and getConfig itself acquires 2 (1 followed by 2)
CheckForUpdates calls onAsyncGet(on a synchronized object) after locking its own mutex. (2 followed by 1)

The fix is to release mutex in CheckForUpdates before calling onAyncGet since that does not have to be in the critical section. Only the read on mtime needs to be under a lock.

Reviewed By: gdrane

Differential Revision: D15457417

fbshipit-source-id: 6f7e8740756b06746c5c6e52e73cee47ccafcb43