Add my_node_id to the Processor
Summary:
- Add the my node ID methods to the processor. (Couldn't add them to the ServerProcessor because unfortunately, there's a lot of stuff in common that depends on it).
- Split the ServerConfig hooks into different hooks based on the functionality.
- Exposed the node_id to the server from the ServerParams and used it to set it in the main ServerProcessor.

NodeID never changes after the server startup. Configs that tries to change it gets rejected, so it's safe to assume that it will be set only once.

For some tests, I passed `NodeID()` as the my_node_id because they weren't setting it in the ServerConfig itself which probably means that it won't hit a codepath that will require reading it (otherwise it would have crashed).

Reviewed By: runemaster

Differential Revision: D14966487

fbshipit-source-id: b8587af4e6e83fa4027b9756db516a8a2d192680