Stagger SHARD_IS_REBULT messages when disable-data-log-rebuilding is true
Summary: Staggering the SHARD_IS_REBUILT messages sent out for the case when disable-data-log-rebuilding is enabled. In this case all the shards with the same max backlog send out the message at exactly the same time, overwhelming the thread 0 on the hosts processing the event log messages. Staggering the messages should help to pace out the processing.

Reviewed By: gdrane

Differential Revision: D15498852

fbshipit-source-id: fad61cf3e449101de6f1571eb9e9f5c46ea89fc7