Nodeset adjustment: combine resizing with randomization
Summary:
If, based on log's throughput, we want to resize nodeset to size n > 127, instead resize it to 127 but randomize it once every retention/min(n/127, 4), so that the union of historical nodesets is around min(n, 127*4) nodes. (Where "4" is a new setting nodeset-max-randomizations.)

The main question was whether and how EpochMetaData should contain information about rate of randomizations. I decided to put unclamped target nodeset size (i.e. can be bigger than 127 and/or cluster size) in EpochMetaData. This way when server starts and activates a sequencer we immediately know n and can calculate the randomization rate and start the timer right away, without waiting for a throughput estimate. A downside is that we need to update epoch store even when target nodeset size changes in a way that doesn't affect nodeset, e.g. from 200 to 500; that's probably not a big deal because such update happens without sequencer reactivation and metadata log append, just a zookeeper roundtrip.

Reviewed By: shobhitdayal

Differential Revision: D14197559

fbshipit-source-id: 698db771b3ee886fddb615479ed00ad7f360fd48