Fix NodeStatsControllerIntegrationTest flakiness
Summary: Under stress, this test times out sometimes and hence is marked as flaky. And most of the time, the test is waiting for a recent version of the config to be available. This diff fixes the flakiness by ensuring that the required config is first written before the server processes are started. This way, on initial load itself, all the processes have the required config.

Reviewed By: MohamedBassem

Differential Revision: D15548537

fbshipit-source-id: cffb7a4e66b2356e9561f2a947dedc198aa7f3f5