Inherit Worker from WorkContext
Summary:
EventLoop which is underlying executor for Worker serially executes requests.
EventLoopTaskQueue is prioritized now, hence task added with priority are
executed in an prioritized manner.

SerialExecutor does not honor priorities, it executes everything enqueued in it
serially. This is not a desired behavior as we want requests, messages and
tasks with different priorities be executed in that manner.

Worker does not need SerialExecutor as today it is based on EventLoop hence
inheriting from WorkContext directly should be fine.

In future this is fine too, if we need the underlying executor to be serial for
example in case of CPUThreadPoolExecutor. We will first wrap the CPUThreadPool
in SerialExecutor and pass in this SerialExecutor to Worker.

Reviewed By: mcrnic, tau0

Differential Revision: D15560481

fbshipit-source-id: ebb6c8611f5ab276c23eb1f064786d040aa3d0b8