Rebuilding filtering: always look for dirty ranges in APPEND data class
Summary:
The UnexpectedNodeOrRackFailureTest fails because mini-rebuilding filters out
records that have the CSI_WRITTEN_BY_REBUILDING flag.

This behavior is incorrect as records with the flag may have their data not
committed to disk even when the flag is set, as described in T43708398 and
https://fb.workplace.com/groups/756141634757724/permalink/830006284037925/.

In this diff I'm just doing the simple fix where we always look at dirty ranges
for the APPEND data class. Justin is working on a more long term fix.

Reviewed By: scsiguy

Differential Revision: D15627310

fbshipit-source-id: 591b91dd171ecf99d2b1db1930751918dad1e802