Augment the maintenances with the safety check results
Summary:
This diff actually fetches the maintenance latest safety check result and augment the responses accordingly.
It also refactors the code the converts Impact to CheckImpactReponse to be generically reusable.

NOTE: In our thrift handlers, we *can* throw exceptions, this is the only place where we allow this to happen since this is the best practice for thrift handlers. folly catches exceptions in continuations of futures and correctly wrap them into the Future object.

Reviewed By: pvjykumar

Differential Revision: D15536658

fbshipit-source-id: 5a7d8c4d870cb677f067a52630aa95e4fee8c6d6