Fewer heap allocations in ProtocolWriter and ProtocolReader
Summary: ProtocolReader/ProtocolWriter is created twice for each message: once to get size, once to actually serialize. It used to allocate a `Source` on the heap, and also make a copy of std::string (dunno how often the string is short enough to avoid heap allocation). It showed up in strobelight profile.

Reviewed By: mcrnic

Differential Revision: D15572328

fbshipit-source-id: 68f6b7caba0c745e2079e2790cceb6bfa18c5232