ZookeeperClientBase: use E::CONNFAILED instead of E::SHUTDOWN for ZCLOSING status
Summary:
In LogDevice E::SHUTDOWN is reserved for logdevice daemon
shutdown and in many callsites people treat it as permanent error
(e.g., give up retries as daemon is shutting down). That's not the
case for ZCLOSING as we recreate zkclients upon zookeeper quorum
changes. Use CONNFAILED as the more appropriate error code.

Reviewed By: shixiao

Differential Revision: D15713387

fbshipit-source-id: dc884233b236c038e1e34ac880e5e9acde500780