Check if socket was closed as part of ReleaseMessage::onReceived.
Summary:
Socket could be closed while the message waits in the queue to be processed. As
part of release message processing we try to get peer node Id. The function
returns invalid nodeID if the socket to node was closed. I
hit a ld_check failure (P65464268) in SealStorageTask creation because of this.

Reviewed By: mcrnic

Differential Revision: D15671107

fbshipit-source-id: db7ddb92847f3862e1e156f39ab48ab148925396