respect source of truth when obtaining NC in initProcessor
Summary:
During our roll out, we could be in a state where NCM is enabled (on the server
side), the source of truth is the existing ServerConfig NC, and the NC in the
NodesConfigurationStore is relatively stale. The code path previously
hard-coded obtaining NC from NCMSource even if the SoT switch is not on, which
caused initProcessor to fail.

For clients, we should be fine since we only enable NCM on clients along with
the seed servers _after_ NCM is enabled on the server (otherwise config
validation fails for the client_settings).

Reviewed By: MohamedBassem

Differential Revision: D15797394

fbshipit-source-id: 27fc06550bba371dc56b598e303cb7c60b559b60