A helper script to read/write NodesConfiguration files
Summary:
A helper binary that you can use to read/write NodesConfiguration files. It can also generate empty NC files (which was the motivation I wrote it at the first place).

The whole idea of the tool is that you read an NC (or generate an empty NC) and then output it in some specific format.

So basically the input is specified with `--path <nc_file>` / `--init` / `--stdin`, the script is smart enough to read json, compressed binary NC, uncomprssed binary NC.
The output is by default json, but you can change that via `--output-format=binary`. You also have the option to output uncompressed binary NC via `--compressed=false`.

Note: The binary format includes the NCWrapper, so you can directly point a FileBasedNCS to it.

Reviewed By: AhmedSoliman

Differential Revision: D15803111

fbshipit-source-id: 38a67a53d2a7ccdfce96ad5aaae9b01a4fe8ec1c