Support for AWAITING_NODES_CONFIG_TRANSITION state instead of RETRY
Summary: It's confusing for the users of the API to see a MaintenanceStatus::RETRY if we are simply waiting for NCM to finish transitions. This diff adds a new explicit state that handle this scenario. The RETRY state will only be used for the exceptional cases like failing to write to the EventLog.

Reviewed By: MohamedBassem

Differential Revision: D15829512

fbshipit-source-id: 0dfcf6d60034cb38ce0925ff314eb02e0aebe614