fix incorrect assert
Summary:
When there's a version mismatch and in debug builds, we check that the read
version from NCS should actually cause a version mismatch, i.e., it is > the
base_version.

We had an erroneous assert where we assert the read version (`stored_version`)
is strictly larger than the new version that we intend to write (which is
base_version + 1).

This diff fixes it.

Reviewed By: runemaster

Differential Revision: D15830553

fbshipit-source-id: 337298df96255a576a463f5dd73592aa42cd6361