Building OSS logdevice with c++17 (#88)
Summary:
Boost ICL library fails to compile on c++17 (https://bugzilla.redhat.com/show_bug.cgi?id=1485641). According to the comments on the bug, there's a new feature in c++17 that's preventing it from compiling and it's controlled by the switch `-fnew-ttp-matching`, so in this diff, i'm disabling it as well.
Pull Request resolved: https://github.com/facebookincubator/LogDevice/pull/88

Differential Revision: D15874102

Pulled By: AhmedSoliman

fbshipit-source-id: aad4d8c1177ec99db7dc6c7a04db3b20cf5011b4