Fix shutdown crash on ClusterMaintenanceState
Summary: stop() of ReplicatedStateMachine instances has to be executed on the worker. In this case the state machine belongs to the Server object. In our case we don't really need to perform stop() because we already stop the processor and workers before this point so we are sure that there is not active work being done. Just destroy it.

Reviewed By: MohamedBassem

Differential Revision: D15897575

fbshipit-source-id: 8b9c191197181a4c77aaa1ae06e6dd9c846d5d45