LifoEventSem::LifoFDBaton suppression for false positive memory leak
Summary:
LSAN does not track ownership correctly through placement new so in some cases LifoFDBaton gets blamed for a memory leak it didn't cause.
Added a test which demonstrates that, before changes, suppressing a leak on an AsyncWaiter will still report a leak on FDBaton which is owned.
After changes it is no longer reported

Reviewed By: yfeldblum

Differential Revision: D15804120

fbshipit-source-id: 0ee79ddb02acaeede693393be8536afd10304f56