LogsDB: apply LD-side write throttling only when LD-side flush policy is enabled
Summary:
The write throttling logic doesn't make much sense when manual flushes are disabled. It's not even using the right setting (memtable-size-per-node vs db-write-buffer-size). I was being dumb and assumed that throttling it's gated and only done if manual flushes are disabled; that wasn't the case.

This diff makes throttling behave the old way when flushes are done the old way.

Along the way it also fixes handling of situation when ld-managed-flushes is true but some other condition for manual flushing is not met - it used to be that (a) RocksDBLogStoreConfig configures rocksdb to do automatic flushes, but PartitionedRocksDBStore does manual flushes too, (b) RocksDBLogStoreConfig didn't make changes to metadata CF options.

Reviewed By: gdrane

Differential Revision: D15818849

fbshipit-source-id: 0b587a1df129de5a0fdfc04f056d427e3b25b6bd