Introduce CompactHistogram and use it for the most numerous histograms
Summary: See comment at the top of Histogram.h and above CompactHistogram. This should save ~1 GB of memory in ServerHistograms alone. I'm also going to use it in RocksDBTablePropertiesCollector to reduce number of buckets and save another 4 GB (most of the saving won't be from the histogram size, it'll just be nicer to implement with the simple power of two buckets).

Reviewed By: shobhitdayal

Differential Revision: D15078358

fbshipit-source-id: f1aa5741cec2300bfc9b2d9737d47bf6c12676cb