Increase default rocksdb-pinned-memtables-limit-percent
Summary:
If `rocksdb-pinned-memtables-limit-percent` is too small relative to `iterator-cache-ttl`, the size of pinned memtable will get over the limit very often, and lots of appends will be rejected. That happened in my test cluster. The default value was too aggressive, this diff increases it to a value that will hopefully be big enough to not worry about it. Instead, we'll need to worry about the high worst-case memory usage, but it has been like that for a long time (before my diff that added `rocksdb-pinned-memtables-limit-percent`), and it didn't cause much trouble.

At some point I'm going to make rocksdb iterators unpin memtables to fix this problem once and for all. Until then, this diff is the best I could think of.

Reviewed By: gdrane

Differential Revision: D15748326

fbshipit-source-id: ffbaa32b84135daee515c52a2c1f09b05cf16fb5