SynchronizationFd crash on usage after fork
Summary:
SynchronizationFD should only be used for signaling within a single process. After a fork it can remain open so operations might be attempted from a child process.
This would start waking up parent and could cause potential stalls in parent or child process.
It is better to abort in this case.q

Reviewed By: tau0

Differential Revision: D16005050

fbshipit-source-id: 72e8d1c6e5e8698391ed3dd05f0276754d9b864a