Set limit on number of messages processed in flight.
Summary:
Earlier all messages were processed inline as soon as they were read from the
event buffer. But with latest changes we push the messages into the per worker
task queue. In case of extreme scenarios, number of messages within queue can
grow large enough to cause the node to OOM. To avoid these OOM and also have
enough messages in the pipeline this diff creates a simple push back mechanism.
When reading every message gets a token, if it cannot get the resource token
the message is not enqueued for further processing. The Socket is scheduled to
read at end of next event loop cycle. If we could get the token, the message is
scheduled for processing. After completion of the message processing the
acquired token is destroyed.

Currently the ResourceBudget is in Processor, if we see higher contention we
can shard and make it better.

Reviewed By: mcrnic

Differential Revision: D15688140

fbshipit-source-id: f1f4329ff53672a4ca1554d834900383d67ce823