Add NodeConfig to NodeState results
Summary:
Since most use cases need to fetch the NodesState and NodesConfig and do the merge on the client side, it would be nice to have the config available in the node state object.

This adds a new `config` property to the `NodeState` object. That means that you only need a single RPC call now to get both the config and state, without doing client-side merging too.

Next Step:
- Update LDOPS 2.0 to perform only get_nodes_state() instead of the two RPC calls.

Reviewed By: MohamedBassem

Differential Revision: D16029256

fbshipit-source-id: 4ad3d46d37ef29ae58587fd189e4e395c2eeda2b