Feature for throttling sequencer reactivations, when possible, and based on the information changed.
Summary:
Context:
Currently, even small changes to the cluster configuration trigger reactivation of almost all the logs. The reactivations trigger recoveries, all near the same time, which can sometimes choke the disk and CPU capacity of the cluster.

Feature:
This feature provides a mechanism to postpone reactivation of logs when it can be safely done. If a reactivation can be postponed, then its new reactivation time is chosen as a random time delay between sequencer-reactivation-min-delay-secs and sequencer-reactivation-max-delay-secs.

Details:
The chosen criterion here, for whether a reactivation can be postponed is as following:

1. Update epoch store immediately but no need to reactivate the sequencer:
This occurs when the nodeset selector choses to keep the old nodeset, and there are no changes to the sequencer options (like window size), and there are no major changes to the logs config (like the replication factor). But there are small changes to the logs config, like the seed, target nodeset size, or the hash has changed.

2. Reactivate the sequencer immediately:
This occurs when any of the important information changes: sequencer options, nodeset selection, replication factor, etc.

3. Reactivate but with a delay (the new feature):
This occurs when neither the sequencer options have changed nor any of the important log config information (like repl factor). More specifically the following conditions must be met:
a.) The nodeset selectors decision is to select a new nodeset but that decision is not due to exclude_from_nodeset flag set on one of the prev nodes.
b.) Neither the sequencer options (like window size), not the log config information (like replication factor) has changed.

Then we infer that the nodeset changed because either the storage state was changed or because the cluster was expanded or shrunk.

PS: It is possible to distinguish whether the storage state was changed or we expanded/shrunk the cluster and provide a different behaviour in terms of reactivation, if that is desirable.

PPS: The many states possible makes the code a little complex. I plan to do some cleanup but I wanted to get it out because it's big and it would be good to get confirmation that the basic idea is sound.

Recommendation for reviewing this diff:
It might be better and simpler to first review the integration test. This way we can be agreed on the requirements. Also have confidence that the diff is passing tests that are testing all the expected functionality. And if some tests are missing then I can work on them right away while the whole diff review is in progress - it'll probably take a while to do.

Reviewed By: runemaster

Differential Revision: D15358563

fbshipit-source-id: 62f058604682b8b7f0c3829731574cf312fa4253