Add timestamps to EpochMetaData (when epoch was incremented, when nodeset was changed).
Summary:
In this diff added two timestamps when epoch was incremented, when nodeset was changed. Also updated `toString` to output those values.

Using set for epoch change to have better readability.

Maybe should be combined with increasing epoch as well to same method, but currently we have different ways to update metadata, so leave that for future improvements.

Reviewed By: al13n321

Differential Revision: D15719645

fbshipit-source-id: 83bd9dba4758f03769d4d6b1b037fb3d1d632237