Fix handling mini-rebuilding in safety checker read availability check
Summary:
Mini rebuilding is a bit awkward because it's the only case at which we consider the shard to be writeable but not readable. Since we now reuse the FailureDomainNodeSet object, we get the artefact of setting these shards with "true" tag while testing for write availability. So, even that we don't want to include these shards in read-availability check, they get included.

NOTE: We don't want to set all other shards to "false" as this operation is expensive so we will do this surgically to avoid safety checker performance regressions.

NOTE: The fix in ldshell is necessary because we currently don't return `ShardDataHealth::LOST_REGIONS` due to historical reasons in our responses to CheckImpact. I will fix that in later diffs.

Reviewed By: MohamedBassem

Differential Revision: D16072626

fbshipit-source-id: 116acd53f9fd9464e3cb5f6c29a9791a488a7eb8