Use NodesConfiguration in copy set selection 1: Copyset selectors
Summary:
As titled. This diff only touches the copyset selector logic
themselves, while the next one will handle actual passing the
switchable NodesConfiguration all the way down to copyset selectors
and get rid of the adaptor constructor still using ServerConfig.

Most of the transitions are mechanical except some tricky part in
WeightedCopySetSelector for weight calculation, as the storage state
in NC is at the per-shard granularity.

Reviewed By: MohamedBassem

Differential Revision: D15981223

fbshipit-source-id: 43f82d558293f51278ff7fb9b14f15b22d0bcda4