Fix flaky test: PartitionedRocksDBStoreTest.PickFlushCandidatesFromOldAndNewPartitions
Summary:
Reason: I was checking the result size to have at least 2 elements there, which is wrong. We generate 4 CFs and from one test run to another the memory size of each CF can be different, for example `91 78 29 10` and `145 28 21 14`. We have target always equals to `108` on the other hand. That means that some of the test runs we can fill it with only first element, but other runs we need more.

To fix: I removed the check on the vector size. We care about the sorting order there, 2 was extra check and that was wrong.

Reviewed By: al13n321

Differential Revision: D16132591

fbshipit-source-id: c4ce8535ed23606de9ebfee3693a5b0d568c01ee